service: Move pkg/svc/svc.go to pkg/service/service.go
Also, eliminate the circular dependency by passing the callback
to service.SyncWithK8s() instead of k8s service frontend values.
