+++ daemon/state.go
@@ -27,6 +27,7 @@ import (
 	"github.com/cilium/cilium/pkg/endpoint/connector"
 	"github.com/cilium/cilium/pkg/k8s"
 	"github.com/cilium/cilium/pkg/labels"
+	"github.com/cilium/cilium/pkg/loadbalancer"
 	"github.com/cilium/cilium/pkg/logging/logfields"
 	"github.com/cilium/cilium/pkg/maps/ctmap"
 	"github.com/cilium/cilium/pkg/maps/lxcmap"
@@ -337,7 +338,10 @@ func (d *Daemon) initRestore(restoredEndpoints *endpointRestoreState) chan struc
 					controller.ControllerParams{
 						DoFunc: func(ctx context.Context) error {
 							frontends := d.k8sSvcCache.UniqueServiceFrontends()
-							return d.svc.SyncWithK8s(frontends)
+							matchSVC := func(addr loadbalancer.L3n4Addr) bool {
+								return frontends.LooseMatch(addr)
+							}
+							return d.svc.SyncWithK8s(matchSVC)
 						},
 					},
 				)
